requirements/requirements-testing.txt
# Pytest for running the tests.
pytest>=5.4.1,<5.5
pytest-django>=3.9.0,<3.10
chardet>=4.0.5
